requirements.txt
Django==3.1.7
djangorestframework==3.12.2
drf_yasg==1.20.0
psycopg2-binary==2.8.6
Pillow==8.1.0
twilio==6.53.0
python-dotenv==0.15.0
django-cors-headers==3.7.0
django-storages==1.11.1
boto3==1.17.17
django-phonenumber-field[phonenumbers]==5.0.0
django-formtools==2.2
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2